setup api service constants
