+++ services/constants/api-paths.ts
@@ -1,0 +1,10 @@
+const nws = {
+  TORNADO_WARNINGS: "",
+  TORNADO_WATCHES: "",
+  STORM_WARNINGS: "",
+  STORM_WATCHES: "",
+};
+
+const swdi = {};
+
+export { nws, swdi };
